pyproject.toml
[project]
name = "localweather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.1",
    "loguru>=0.7.3",
    "luigi>=3.6.0",
    "openmeteo-requests>=1.3.0",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.0.1",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "setuptools>=77.0.3",
    "streamlit>=1.43.2",
    "tool>=0.8.0",
]
